refactor: split endCombatPhaseCommand.ts into focused modules
Extract damage resolution and summoned enemy handling into separate files:
- damageResolution.ts: handles pending damage calculation, fame/reputation changes
- summonedEnemyHandling.ts: handles summon abilities and discarding summoned enemies
- endCombatPhaseCommand.ts: now contains only phase transition and combat end logic

Reduces original 1,038 line file to ~583 lines with better separation of concerns.
